+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for mypy-json-report
 
+## Unreleased
+
+- Use GA version of Python 3.11 in test matrix.
+
 ## v0.1.3 [2022-09-07]
 
 - Removed upper bound for Python compatibility